Se completa la funcionalidad para crear usuario para asesor y usuario para cliente manejando una logica diferente para cada uno. También se hace la notificación respectiva para cada uno
