+++ src/models/user.model.ts
@@ -56,9 +56,8 @@ export class User extends Entity {
 
   @property({
     type: 'string',
-    required: true,
   })
-  password: string;
+  password?: string;
 
   @property({
     type: 'string',
@@ -68,7 +67,7 @@ export class User extends Entity {
   @property({
     type: 'boolean',
   })
-  hashState: boolean;
+  hashState?: boolean;
 
   @property({
     type: 'number',
